kontrol/{utils,state_record,foundry,__main__}: separate out foundry_state_load into its own module
