+++ src/kontrol/utils.py
@@ -107,6 +107,12 @@ def kontrol_up_to_date(digest_file: Path) -> bool:
     return digest_dict.get('kontrol', '') == VERSION
 
 
+def read_contract_names(contract_names: Path) -> dict[str, str]:
+    if not contract_names.exists():
+        raise FileNotFoundError(f'Contract names dictionary file not found: {contract_names}')
+    return json.loads(contract_names.read_text())
+
+
 def parse_test_version_tuple(value: str) -> tuple[str, int | None]:
     if ':' in value:
         test, version = value.split(':')
